requirements.txt
nicegui
selenium
pandas
python-dotenv
openpyxl
SpeechRecognition
requests
openai
